worker/requirements.txt
# Worker dependencies
# Shared with backend but minimal set for generation tasks

# Database
psycopg2-binary>=2.9.9
duckdb>=0.9.0

# Gemini AI
google-genai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
faker>=20.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
